genlang: use git for release string retrieval.
Retrieve the strings for release versions from git similar as done for new git
versions. This fixes missing support for releases after 3.10. Furthermore older
releases now also use git.

Since it might be acceptable to not support older daily / current builds
anymore the svn server is not absolutely necessary anymore. Note that there
might still be a lot of pre-git current builds in use, so keeping the svn
server around for a while would be good. Adjusting the script to use the svn
information in the git repository should be possible as well but might require
access to a local clone.

Change-Id: I7f1595453ebb6de555e1272740128699a3cc8df5