Update patch set 2
Patch Set 2:

Does querying the database for the results in rbmaster.pm work at all? I've only had a short look at it but from my understanding the database is updated with the build results by checklog.pl, which by itself is called by hooks/buildcomplete.sh. This indicates to me that it's not called from the buildmaster, thus reading the results from the database in the buildmaster is likely to be done before the new values are available (rbmaster.pm updates the database but not the errors and warnings fields).

I might be completely wrong with my understanding here though :)

Patch-set: 2
Reviewer: Gerrit User 1000001 <1000001@583131e1-1a33-4f2b-b8ec-82434880f207>
Label: Code-Review=0
Label: Tested=0