builds:  Add erosqnative, iaudio x3/m3/m5, cowond2 manual builds
(iriverihp120 and ipod mini2g were already part of the set)

  Crucially, also have them upload results.

This is the first step towards pushing manual builds out to the farm
instead of a nightly run on the main server.

Use this set to work out the kinks on the server side.  Once
that's solid, then expand to the full set of manual targets and
redo the nightly manual run to assemble then last build
instead of redoing everything.

Change-Id: I5122900de9955a09289b6a8c766d50cf687bda1c